# api/api.yaml
openapi: 3.0.3

info:
  version: 1.0.0
  title: Inventario de aplicaciones GSS
  description: Api REST del inventario de aplicaciones GSS
  license:
    name: MIT
    url: 'https://opensource.org/licenses/MIT'
  contact:
    email: jesvizan@jcyl.es
  termsOfService: 'https://www.jcyl.es/terminosServicio'
servers:
  - url: 'https://jcweb24des101.ae.jcyl.es/{basePath}'
    variables:
      basePath:
        default: /api/v1
paths:
  /sistemas:
    get:
      summary: Listado paginado de sistemas de informacón
      operationId: getSistemasInformacion
      tags:
        - sistemas
      parameters:
        - name: page
          in: query
          required: false
        - name: page_size
          in: query
          required: false
      responses:
        '200':
          description: Operación realizada con éxito
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ListadoPaginadoResumenSistemas'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        default:
          $ref: '#/components/responses/UnspecifiedError'
    post:
      summary: Registro de un sistema de información
      operationId: registrarSistemaInformacion
      tags:
        - sistemas
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SistemaInformacion'
      responses:
        '200':
          description: Operación realizada con éxito
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SistemaInformacion'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        default:
          $ref: '#/components/responses/UnspecifiedError'

  /sistemas/{sistemaId}:
    get:
      summary: Recuperar la información de un sistema
      operationId: getSistemaInformacion
      tags:
        - sistemas
      parameters:
        - name: sistemaId
          in: path
          required: true
          allowEmptyValue: false
          schema:
            type: string

      responses:
        '200':
          description: Operación realizada con éxito
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SistemaInformacion'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        default:
          $ref: '#/components/responses/UnspecifiedError'

    put:
      summary: Modificar la información de un sistema
      operationId: updateSistemaInformacion
      tags:
        - sistemas
      parameters:
        - name: sistemaId
          in: path
          required: true
          allowEmptyValue: false
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SistemaInformacion'

      responses:
        '200':
          description: Operación realizada con éxito
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SistemaInformacion'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        default:
          $ref: '#/components/responses/UnspecifiedError'

  /sistemas/{sistemaId}/componentes:
    get:
      summary: Recuperar los componentes de un sistema
      operationId: getComponentesSistemaInformacion
      tags:
        - sistemas
      parameters:
        - name: sistemaId
          in: path
          required: true
          allowEmptyValue: false
          schema:
            type: string

      responses:
        '200':
          description: Operación realizada con éxito
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ComponenteSistema'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        default:
          $ref: '#/components/responses/UnspecifiedError'

  /sistemas/{sistemaId}/componentes/{componenteId}:
    get:
      summary: Recuperar un componente del sistema
      operationId: getComponenteSistemaInformacion
      tags:
        - sistemas
      parameters:
        - name: sistemaId
          in: path
          required: true
          allowEmptyValue: false
          schema:
            type: string
        - name: componenteId
          in: path
          required: true
          allowEmptyValue: false
          schema:
            type: string

      responses:
        '200':
          description: Operación realizada con éxito
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ComponenteSistema'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        default:
          $ref: '#/components/responses/UnspecifiedError'

  /dir3/unidades:
    get:
      summary: consulta de unidades dir3
      operationId: queryUnidadesDIR3
      tags:
        - dir3
      parameters:
        - name: nombre
          in: query
          required: false
          schema:
            type: string
        - name: id
          in: query
          required: false
          schema:
            type: string
      responses:
        '200':
          description: Operación realizada con éxito
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ListadoPaginadoUnidadesDIR3'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        default:
          $ref: '#/components/responses/UnspecifiedError'

  /dir3/unidades/{unidad_id}:
    get:
      summary: Recuperar unidad DIR3
      operationId: getUnidadDIR3
      tags:
        - dir3
      parameters:
        - name: unidad_id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Operación realizada con éxito
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UnidadDIR3'
        '404':
          $ref: '#/components/responses/NotFound'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        default:
          $ref: '#/components/responses/UnspecifiedError'

components:
  securitySchemes:
    keycloak:
      type: oauth2
      flows:
        authorizationCode:
          authorizationUrl: >-
            https://ssokcpre.jcyl.es/auth/realms/GSS-conciliacion-extranet/protocol/openid-connect/auth
          tokenUrl: >-
            https://ssokcpre.jcyl.es/auth/realms/GSS-conciliacion-extranet/protocol/openid-connect/token
          scopes:
            peticiones_write: Modificación de peticiones
            convocatorias_read: Consulta de convocatorias
            convocatorias_write: Modificación de convocatorias
            centros_read: Consulta de centros
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  responses:
    DummyResponse:
      description: Respuesta para generar stubs
      content:
        application/json:
          schema:
            type: object
    Void:
      description: Respuesta sin contenido
      content:
        application/json:
          schema:
            type: object

    Unauthorized:
      description: Debe indicar datos de autorización en la petición
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Forbidden:
      description: No tiene permiso para ejecutar la operación
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    UnspecifiedError:
      description: Error inesperado
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    NotFound:
      description: El elemento solicitado no existe
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'

  schemas:
    Error:
      type: object
      required:
        - codigo
        - descripcion
      properties:
        codigo:
          type: string
        descripcion:
          type: string
        descripcionDetallada:
          type: string

    ListadoPaginadoResumenSistemas:
      type: object
      required:
        - total
        - page
        - page_size
        - items
      properties:
        total:
          type: integer
        page:
          type: integer
        page_size:
          type: integer
        items:
          type: array
          items:
            $ref: '#/components/schemas/ResumenSistemaInformacion'

    ResumenSistemaInformacion:
      type: object
      required:
        - sistema_id
        - nombre
      properties:
        sistema_id:
          type: string
        nombre:
          type: string
        unidad_responsable:
          type: string
        tecnico_responsable:
          type: string

    SistemaInformacion:
      type: object
      required:
        - sistema_id
        - nombre
      properties:
        sistema_id:
          type: string
        nombre:
          type: string
        unidad_responsable:
          type: string
        tecnico_responsable:
          type: string
        observaciones:
          type: string
#        componentes:
#          type: array
#          items:
#            $ref: "#/components/schemas/ComponenteSistema"

    UnidadDIR3:
      type: object
      required:
        - unidad_id
        - nombre
      properties:
        unidad_id:
          type: string
        unidad_padre_id:
          type: string
        nombre:
          type: string
        nombre_unidad_padre:
          type: string

    ListadoPaginadoUnidadesDIR3:
      type: object
      required:
        - total
        - page
        - page_size
        - items
      properties:
        total:
          type: integer
        page:
          type: integer
        page_size:
          type: integer
        items:
          type: array
          items:
            $ref: '#/components/schemas/UnidadDIR3'

    ComponenteSistema:
      type: object
      required:
        - sistema_id
        - componente_id
        - tipo
        - nombre
      properties:
        sistema_id:
          type: string
        tipo:
          type: string
        componente_id:
          type: string
        nombre:
          type: string
        git_repo:
          type: string
        observaciones:
          type: string